refactor: reorganize chain modules into focused, domain-specific files
- Split monolithic subspace.ts into modular structure:
  - balance.ts: System and balances pallet queries
  - emission0.ts: Emission pallet functions
  - torus0/agents.ts: Agent registration and management
  - torus0/namespace.ts: Namespace operations
  - torus0/staking.ts: Staking and stake queries
- Move date/time utilities to @packages/utils/src/date-time.ts
- Consolidate timestamp functions into single nowISOString()
- Reorganize result utilities into result/ directory structure
- Update all imports to use new modular paths
- Rename queryBurnValue to queryAgentBurn for clarity
- Update chain/CLAUDE.md documentation to reflect new structure
