+++ packages/torus-sdk-ts/src/chain/index.ts
@@ -1,6 +1,7 @@
 export * from "./common/index.js";
 
-export * from "./subspace.js";
+export * from "./balance.js";
 export * from "./governance.js";
-export * from "./torus0.js";
+export * from "./torus0/index.js";
+export * from "./emission0.js";
 export * from "./permission0.js";
